# remove unneed hashcount, add xxh3 8.33% faster