--- a/go.mod
+++ b/go.mod
@@ -4,5 +4,8 @@ go 1.23
 
 require (
 	github.com/klauspost/compress v1.16.7
+	github.com/zeebo/xxh3 v1.0.2
 	go.mongodb.org/mongo-driver v1.17.2
 )
+
+require github.com/klauspost/cpuid/v2 v2.0.9 // indirect