Per the name of the branch.  Initially tried to incorporate roaring bitmaps.  Was about 30% slower.  This is down to matching the current code.  If I did batch compares it might be faster but would look much different.  

However, removed `HashCount`.  In general use range as it allows the compiler to make optimizations.
Switch from FNV much higher quality (and faster xxh3).

```bash
goos: linux
goarch: amd64
pkg: github.com/starskey-io/starskey/bloomfilter
cpu: AMD Ryzen 9 6900HX with Radeon Graphics        
            │   old.txt   │               new.txt               │
            │   sec/op    │   sec/op     vs base                │
Add-16        69.74n ± 0%   66.99n ± 1%   -3.94% (p=0.000 n=25)
Contains-16   46.60n ± 1%   40.77n ± 1%  -12.51% (p=0.000 n=25)
geomean       57.01n        52.26n        -8.33%

            │   old.txt    │               new.txt               │
            │     B/op     │    B/op     vs base                 │
Add-16        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=25) ¹
Contains-16   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=25) ¹
geomean                  ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

            │   old.txt    │               new.txt               │
            │  allocs/op   │ allocs/op   vs base                 │
Add-16        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=25) ¹
Contains-16   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=25) ¹
geomean                  ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```